backend/requirements.txt
fastapi
uvicorn
langchain
pymongo
sqlalchemy
python-dotenv
pydantic
pydantic-settings
